Preliminary support for 1.20.4+ and NeoForge
Notably absent are mod integrations with Hourglass and Snow Real Magic,
both currently due to their lack of support for some or all of the new
versions that Immersive Snow targets.

While everything seems to compile and work fine, there may still be
potential bugs and errors (e.g. possibly something relating to sea level).

Also fixed a potential bug where it was possible that snow in the Nether
would never melt if a season mod was installed.