Fix performance issues, once and for all.
Turns out the performance issues were a bit multi-faceted -- block
updates were trying to update blocks in unloaded chunks, which would try
to load them. So, a check was added to ensure neighboring chunks are
loaded. Then, the queue system was revamped into two queues -- for the
current tick, and the next tick. A mixin to the chunk random tick was
added that sends chunks straight to the current tick queue, whilst
chunks received via the load event are held until the next tick -- this
avoids a deadlock due to trying to access the chunk before it is
promoted to full status, which was causing either massive slowdowns or
incredible lag, or both.

Hopefully there is no more performance issues after this. Hopefully.